backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pydantic
pytest
httpx
